Refactor strokes data into a separate class
